Updated node size logic
I have reason to believe it is faulty, as the sizes are incorrectly calculated in some places but not others. This causes the rankOf function to fail. I haven't yet found out where, but it appears that there are only problems on the left side of the root node thus far.